+++ projects/ng-qrcode-demo/src/app/app.component.ts
@@ -1,4 +1,5 @@
 import { Component } from "@angular/core"
+import { QrCodeErrorCorrectionLevel } from "ng-qrcode"
 
 @Component({
   selector: "app-root",
@@ -9,9 +10,8 @@ export class AppComponent {
 
   value = "https://www.npmjs.com/package/ng-qrcode"
   size = "300"
-  errorCorrectionLevel = "M"
-  // TODO refactor to QRCodeErrorCorrectionLevel after https://github.com/DefinitelyTyped/DefinitelyTyped/pull/35367
-  errorCorrectionLevels = ["L", "M", "Q", "H"]
+  errorCorrectionLevel: QrCodeErrorCorrectionLevel = "M"
+  errorCorrectionLevels: QrCodeErrorCorrectionLevel[] = ["L", "M", "Q", "H"]
 
   get example() {
     return `<qr-code value="${this.value}" size="${this.size}" errorCorrectionLevel="${this.errorCorrectionLevel}"></qr-code>`