+++ projects/ng-qrcode/src/lib/types.ts
@@ -1,0 +1,1 @@
+export type QrCodeErrorCorrectionLevel = "low" | "medium" | "quartile" | "high" | "L" | "M" | "Q" | "H"